+++ pineappl_py/src/evolution.rs
@@ -1,10 +1,64 @@
+//! Evolution interface.
+
+use super::pids::PyPidBasis;
 use numpy::{IntoPyArray, PyArray1};
 use pineappl::evolution::{EvolveInfo, OperatorSliceInfo};
-
 use pyo3::prelude::*;
 
+/// PyO3 wrapper to :rustdoc:`pineappl::evolution::OperatorSliceInfo <evolution/struct.OperatorSliceInfo.html>`.
+#[pyclass(name = "OperatorSliceInfo")]
+#[derive(Clone)]
+#[repr(transparent)]
+pub struct PyOperatorSliceInfo {
+    pub(crate) info: OperatorSliceInfo,
+}
+
+#[pymethods]
+impl PyOperatorSliceInfo {
+    /// Constructor.
+    ///
+    /// Parameteters
+    /// ------------
+    /// fac0 : float
+    ///     initial factorization scale
+    /// pids0 : list(int)
+    ///     flavors available at the initial scale
+    /// x0 : list(float)
+    ///     x-grid at the initial scale
+    /// fac1 : float
+    ///     evolved final scale
+    /// pids1 : list(int)
+    ///     flavors available at the final scale
+    /// x1 : list(float)
+    ///     x-grid at the final scale
+    /// pid_basis : PyPidBasis
+    ///     flavor basis reprentation at the initial scale
+    #[new]
+    pub fn new(
+        fac0: f64,
+        pids0: Vec<i32>,
+        x0: Vec<f64>,
+        fac1: f64,
+        pids1: Vec<i32>,
+        x1: Vec<f64>,
+        pid_basis: PyPidBasis,
+    ) -> Self {
+        Self {
+            info: OperatorSliceInfo {
+                fac0,
+                pids0,
+                x0,
+                fac1,
+                pids1,
+                x1,
+                pid_basis: pid_basis.into(),
+            },
+        }
+    }
+}
+
 /// PyO3 wrapper to :rustdoc:`pineappl::evolution::EvolveInfo <evolution/struct.EvolveInfo.html>`.
-#[pyclass]
+#[pyclass(name = "EvolveInfo")]
 #[repr(transparent)]
 pub struct PyEvolveInfo {
     pub(crate) evolve_info: EvolveInfo,
@@ -37,10 +91,16 @@ impl PyEvolveInfo {
     }
 }
 
-/// PyO3 wrapper to :rustdoc:`pineappl::evolution::OperatorSliceInfo <evolution/struct.OperatorSliceInfo.html>`.
-#[pyclass]
-#[repr(transparent)]
-#[derive(Clone)]
-pub struct PyOperatorSliceInfo {
-    pub(crate) slice_info: OperatorSliceInfo,
+/// Register submodule in parent.
+pub fn register(parent_module: &Bound<'_, PyModule>) -> PyResult<()> {
+    let m = PyModule::new_bound(parent_module.py(), "evolution")?;
+    m.setattr(pyo3::intern!(m.py(), "__doc__"), "Evolution interface.")?;
+    pyo3::py_run!(
+        parent_module.py(),
+        m,
+        "import sys; sys.modules['pineappl.evolution'] = m"
+    );
+    m.add_class::<PyEvolveInfo>()?;
+    m.add_class::<PyOperatorSliceInfo>()?;
+    parent_module.add_submodule(&m)
 }