+++ pineappl_py/src/grid.rs
@@ -1,196 +1,30 @@
-use ndarray::CowArray;
-use pineappl::boc::Order;
-use pineappl::convolutions::LumiCache;
-use pineappl::evolution::{AlphasTable, OperatorInfo, OperatorSliceInfo};
-use pineappl::grid::{Grid, Ntuple};
-use pineappl::pids::PidBasis;
+//! Grid interface.
 
 use super::bin::PyBinRemapper;
-use super::channel::PyChannel;
-use super::evolution::PyEvolveInfo;
+use super::boc::{PyChannel, PyOrder};
+use super::evolution::{PyEvolveInfo, PyOperatorSliceInfo};
 use super::fk_table::PyFkTable;
 use super::subgrid::{PySubgridEnum, PySubgridParams};
-
 use itertools::izip;
-use numpy::{
-    IntoPyArray, PyArray1, PyArrayMethods, PyReadonlyArray1, PyReadonlyArray4, PyReadonlyArray5,
-};
-
+use ndarray::CowArray;
+use numpy::{IntoPyArray, PyArray1, PyArrayMethods, PyReadonlyArray1, PyReadonlyArray4};
+use pineappl::convolutions::LumiCache;
+use pineappl::evolution::AlphasTable;
+use pineappl::grid::{Grid, Ntuple};
+use pyo3::exceptions::PyValueError;
+use pyo3::prelude::*;
+use pyo3::types::PyIterator;
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::BufReader;
 use std::path::PathBuf;
 
-use pyo3::exceptions::PyValueError;
-use pyo3::prelude::*;
-use pyo3::types::PyIterator;
-
-/// PyO3 wrapper to :rustdoc:`pineappl::grid::Order <grid/struct.Order.html>`.
-#[pyclass]
-#[repr(transparent)]
-pub struct PyOrder {
-    pub(crate) order: Order,
-}
-
-// TODO: should probably be in a different module
-// TODO: rename to `PidBasis`
-#[pyclass]
-#[derive(Clone)]
-pub enum PyPidBasis {
-    Pdg,
-    Evol,
-}
-
-impl From<PyPidBasis> for PidBasis {
-    fn from(basis: PyPidBasis) -> Self {
-        match basis {
-            PyPidBasis::Pdg => Self::Pdg,
-            PyPidBasis::Evol => Self::Evol,
-        }
-    }
-}
-
-// TODO: should probably be in a different module
-// TODO: rename to `OperatorSliceInfo`
-#[pyclass]
-#[derive(Clone)]
-pub struct PyOperatorSliceInfo {
-    info: OperatorSliceInfo,
-}
-
-#[pymethods]
-impl PyOperatorSliceInfo {
-    /// Constructor.
-    ///
-    /// Parameteters
-    /// ------------
-    /// fac0 : float
-    ///     initial factorization scale
-    /// pids0 : list(int)
-    ///     flavors available at the initial scale
-    /// x0 : list(float)
-    ///     x-grid at the initial scale
-    /// fac1 : float
-    ///     evolved final scale
-    /// pids1 : list(int)
-    ///     flavors available at the final scale
-    /// x1 : list(float)
-    ///     x-grid at the final scale
-    /// pid_basis : PyPidBasis
-    ///     flavor basis reprentation at the initial scale
-    #[new]
-    pub fn new(
-        fac0: f64,
-        pids0: Vec<i32>,
-        x0: Vec<f64>,
-        fac1: f64,
-        pids1: Vec<i32>,
-        x1: Vec<f64>,
-        pid_basis: PyPidBasis,
-    ) -> Self {
-        Self {
-            info: OperatorSliceInfo {
-                fac0,
-                pids0,
-                x0,
-                fac1,
-                pids1,
-                x1,
-                pid_basis: pid_basis.into(),
-            },
-        }
-    }
-}
-
-impl PyOrder {
-    pub(crate) fn new(order: Order) -> Self {
-        Self { order }
-    }
-}
-
-#[pymethods]
-impl PyOrder {
-    /// Constructor.
-    ///
-    /// Parameters
-    /// ----------
-    /// alphas : int
-    ///     power of :math:`\alpha_s`
-    /// alpha : int
-    ///     power of :math:`\alpha`
-    /// logxir : int
-    ///     power of :math:`\ln(\xi_r)`
-    /// logxif : int
-    ///     power of :math:`\ln(\xi_f)`
-    #[new]
-    pub fn new_order(alphas: u32, alpha: u32, logxir: u32, logxif: u32) -> Self {
-        Self::new(Order::new(alphas, alpha, logxir, logxif))
-    }
-
-    /// Tuple representation.
-    ///
-    /// Returns
-    /// -------
-    /// alphas : int
-    ///     power of :math:`\alpha_s`
-    /// alpha : int
-    ///     power of :math:`\alpha`
-    /// logxir : int
-    ///     power of :math:`\ln(\xi_r)`
-    /// logxif : int
-    ///     power of :math:`\ln(\xi_f)`
-    pub fn as_tuple(&self) -> (u32, u32, u32, u32) {
-        (
-            self.order.alphas,
-            self.order.alpha,
-            self.order.logxir,
-            self.order.logxif,
-        )
-    }
-
-    /// Return a mask suitable to pass as the `order_mask` parameter of [`Grid::convolve`].
-    ///
-    /// The selection of `orders` is controlled using the `max_as` and `max_al` parameters, for
-    /// instance setting `max_as = 1` and `max_al = 0` selects the LO QCD only, `max_as = 2` and
-    /// `max_al = 0` the NLO QCD; setting `max_as = 3` and `max_al = 2` would select all NLOs, and
-    /// the NNLO QCD.
-    ///
-    /// See `pineappl` crate docs for more examples.
-    ///
-    /// Returns
-    /// -------
-    /// numpy.ndarray(bool)
-    ///     boolean array, to be used as orders' mask
-    #[staticmethod]
-    pub fn create_mask<'py>(
-        orders: Vec<PyRef<Self>>,
-        max_as: u32,
-        max_al: u32,
-        logs: bool,
-        py: Python<'py>,
-    ) -> Bound<'py, PyArray1<bool>> {
-        Order::create_mask(
-            &orders.iter().map(|o| o.order.clone()).collect::<Vec<_>>(),
-            max_as,
-            max_al,
-            logs,
-        )
-        .into_pyarray_bound(py)
-    }
-}
-
 /// PyO3 wrapper to :rustdoc:`pineappl::grid::Grid <grid/struct.Grid.html>`.
-#[pyclass]
+#[pyclass(name = "Grid")]
 #[repr(transparent)]
 #[derive(Clone)]
 pub struct PyGrid {
     pub(crate) grid: Grid,
-}
-
-impl PyGrid {
-    pub(crate) fn new(grid: Grid) -> Self {
-        Self { grid }
-    }
 }
 
 #[pymethods]
@@ -214,12 +48,14 @@ impl PyGrid {
         bin_limits: PyReadonlyArray1<f64>,
         subgrid_params: PySubgridParams,
     ) -> Self {
-        Self::new(Grid::new(
-            channels.iter().map(|pyc| pyc.entry.clone()).collect(),
-            orders.iter().map(|pyo| pyo.order.clone()).collect(),
-            bin_limits.to_vec().unwrap(),
-            subgrid_params.subgrid_params,
-        ))
+        Self {
+            grid: Grid::new(
+                channels.iter().map(|pyc| pyc.entry.clone()).collect(),
+                orders.iter().map(|pyo| pyo.order.clone()).collect(),
+                bin_limits.to_vec().unwrap(),
+                subgrid_params.subgrid_params,
+            ),
+        }
     }
 
     /// Add a point to the grid.
@@ -425,15 +261,16 @@ impl PyGrid {
     /// numpy.ndarray(float) :
     ///     cross sections for all bins, for each scale-variation tuple (first all bins, then
     ///     the scale variation)
+    #[pyo3(signature = (pdg_id, xfx, alphas, order_mask = None, bin_indices = None, channel_mask = None, xi = None))]
     pub fn convolve_with_one<'py>(
         &self,
         pdg_id: i32,
         xfx: &Bound<'py, PyAny>,
         alphas: &Bound<'py, PyAny>,
-        order_mask: PyReadonlyArray1<bool>,
-        bin_indices: PyReadonlyArray1<usize>,
-        channel_mask: PyReadonlyArray1<bool>,
-        xi: Vec<(f64, f64)>,
+        order_mask: Option<PyReadonlyArray1<bool>>,
+        bin_indices: Option<PyReadonlyArray1<usize>>,
+        channel_mask: Option<PyReadonlyArray1<bool>>,
+        xi: Option<Vec<(f64, f64)>>,
         py: Python<'py>,
     ) -> Bound<'py, PyArray1<f64>> {
         let mut xfx = |id, x, q2| xfx.call1((id, x, q2)).unwrap().extract().unwrap();
@@ -443,10 +280,10 @@ impl PyGrid {
         self.grid
             .convolve(
                 &mut lumi_cache,
-                &order_mask.to_vec().unwrap(),
-                &bin_indices.to_vec().unwrap(),
-                &channel_mask.to_vec().unwrap(),
-                &xi,
+                &order_mask.map_or(vec![], |b| b.to_vec().unwrap()),
+                &bin_indices.map_or(vec![], |c| c.to_vec().unwrap()),
+                &channel_mask.map_or(vec![], |d| d.to_vec().unwrap()),
+                &xi.map_or(vec![(1.0, 1.0)], |m| m),
             )
             .into_pyarray_bound(py)
     }
@@ -487,6 +324,7 @@ impl PyGrid {
     ///     numpy.ndarray(float) :
     ///         cross sections for all bins, for each scale-variation tuple (first all bins, then
     ///         the scale variation)
+    #[pyo3(signature = (pdg_id1, xfx1, pdg_id2, xfx2, alphas, order_mask = None, bin_indices = None, channel_mask = None, xi = None))]
     pub fn convolve_with_two<'py>(
         &self,
         pdg_id1: i32,
@@ -494,10 +332,10 @@ impl PyGrid {
         pdg_id2: i32,
         xfx2: &Bound<'py, PyAny>,
         alphas: &Bound<'py, PyAny>,
-        order_mask: PyReadonlyArray1<bool>,
-        bin_indices: PyReadonlyArray1<usize>,
-        channel_mask: PyReadonlyArray1<bool>,
-        xi: Vec<(f64, f64)>,
+        order_mask: Option<PyReadonlyArray1<bool>>,
+        bin_indices: Option<PyReadonlyArray1<usize>>,
+        channel_mask: Option<PyReadonlyArray1<bool>>,
+        xi: Option<Vec<(f64, f64)>>,
         py: Python<'py>,
     ) -> Bound<'py, PyArray1<f64>> {
         let mut xfx1 = |id, x, q2| xfx1.call1((id, x, q2)).unwrap().extract().unwrap();
@@ -509,88 +347,12 @@ impl PyGrid {
         self.grid
             .convolve(
                 &mut lumi_cache,
-                &order_mask.to_vec().unwrap(),
-                &bin_indices.to_vec().unwrap(),
-                &channel_mask.to_vec().unwrap(),
-                &xi,
+                &order_mask.map_or(vec![], |b| b.to_vec().unwrap()),
+                &bin_indices.map_or(vec![], |c| c.to_vec().unwrap()),
+                &channel_mask.map_or(vec![], |d| d.to_vec().unwrap()),
+                &xi.map_or(vec![(1.0, 1.0)], |m| m),
             )
             .into_pyarray_bound(py)
-    }
-
-    /// Convolve with an evolution operator.
-    ///
-    /// Parameters
-    /// ----------
-    /// operator : numpy.ndarray(int, rank=5)
-    ///     evolution tensor
-    /// fac0 : float
-    ///     reference scale
-    /// pids0 : numpy.ndarray(int)
-    ///     sorting of the particles in the tensor for final FkTable
-    /// x0 : numpy.ndarray(float)
-    ///     final FKTable interpolation grid
-    /// fac1 : numpy.ndarray(float)
-    ///     list of factorization scales
-    /// pids1 : numpy.ndarray(int)
-    ///     sorting of the particles in the grid
-    /// x1 : numpy.ndarray(float)
-    ///     interpolation grid at process level
-    /// ren1 : numpy.ndarray(float)
-    ///     list of renormalization scales
-    /// alphas : numpy.ndarray(float)
-    ///     list with :math:`\alpha_s(Q2)` for the process scales
-    /// xi : (float, float)
-    ///     factorization and renormalization variation
-    /// pid_basis : str
-    ///     type of channel identifier
-    /// order_mask : numpy.ndarray(bool)
-    ///     boolean mask to activate orders
-    ///
-    /// Returns
-    /// -------
-    /// PyFkTable :
-    ///     produced FK table
-    #[deprecated(since = "0.7.4", note = "use evolve_with_slice_iter instead")]
-    pub fn evolve(
-        &self,
-        operator: PyReadonlyArray5<f64>,
-        fac0: f64,
-        pids0: PyReadonlyArray1<i32>,
-        x0: PyReadonlyArray1<f64>,
-        fac1: PyReadonlyArray1<f64>,
-        pids1: PyReadonlyArray1<i32>,
-        x1: PyReadonlyArray1<f64>,
-        ren1: PyReadonlyArray1<f64>,
-        alphas: PyReadonlyArray1<f64>,
-        xi: (f64, f64),
-        pid_basis: String,
-        order_mask: PyReadonlyArray1<bool>,
-    ) -> PyFkTable {
-        let op_info = OperatorInfo {
-            fac0: fac0,
-            pids0: pids0.to_vec().unwrap(),
-            x0: x0.to_vec().unwrap(),
-            fac1: fac1.to_vec().unwrap(),
-            pids1: pids1.to_vec().unwrap(),
-            x1: x1.to_vec().unwrap(),
-            ren1: ren1.to_vec().unwrap(),
-            alphas: alphas.to_vec().unwrap(),
-            xir: xi.0,
-            xif: xi.1,
-            pid_basis: pid_basis.parse().unwrap(),
-        };
-
-        let evolved_grid = self
-            .grid
-            .evolve(
-                operator.as_array(),
-                &op_info,
-                order_mask.as_slice().unwrap(),
-            )
-            .expect("Nothing returned from evolution.");
-        PyFkTable {
-            fk_table: evolved_grid,
-        }
     }
 
     /// Collect information for convolution with an evolution operator.
@@ -739,7 +501,9 @@ impl PyGrid {
     ///     grid
     #[staticmethod]
     pub fn read(path: PathBuf) -> Self {
-        Self::new(Grid::read(BufReader::new(File::open(path).unwrap())).unwrap())
+        Self {
+            grid: Grid::read(BufReader::new(File::open(path).unwrap())).unwrap(),
+        }
     }
 
     /// Write to file.
@@ -770,20 +534,6 @@ impl PyGrid {
     /// Merge with another grid.
     pub fn merge(&mut self, other: Self) -> PyResult<()> {
         match self.grid.merge(other.grid) {
-            Ok(()) => Ok(()),
-            Err(x) => Err(PyValueError::new_err(format!("{:?}", x))),
-        }
-    }
-
-    /// Merge with another grid, loaded from file.
-    ///
-    /// Note
-    /// ----
-    /// For a current limitation with the implementation of the bound object `Grid` is not possible
-    /// to operate with two `Grid`s in memory, since is not possible to pass a `Grid` by argument
-    #[deprecated = "Deprecated in favor of PyGrid::merge"]
-    pub fn merge_from_file(&mut self, path: PathBuf) -> PyResult<()> {
-        match self.grid.merge(Self::read(path).grid) {
             Ok(()) => Ok(()),
             Err(x) => Err(PyValueError::new_err(format!("{:?}", x))),
         }
@@ -913,4 +663,18 @@ impl PyGrid {
     pub fn delete_bins(&mut self, bin_indices: PyReadonlyArray1<usize>) {
         self.grid.delete_bins(&bin_indices.to_vec().unwrap())
     }
+}
+
+/// Register submodule in parent.
+pub fn register(parent_module: &Bound<'_, PyModule>) -> PyResult<()> {
+    let m = PyModule::new_bound(parent_module.py(), "grid")?;
+    m.setattr(pyo3::intern!(m.py(), "__doc__"), "Grid interface.")?;
+    pyo3::py_run!(
+        parent_module.py(),
+        m,
+        "import sys; sys.modules['pineappl.grid'] = m"
+    );
+    m.add_class::<PyGrid>()?;
+    m.add_class::<PyOrder>()?;
+    parent_module.add_submodule(&m)
 }