+++ pineappl_py/src/lib.rs
@@ -1,35 +1,29 @@
-// this is needed for PyO3 to work
+//! Generate PyO3 interface.
+
 #![allow(unsafe_op_in_unsafe_fn)]
 
 use pyo3::prelude::*;
 
 pub mod bin;
-pub mod channel;
+pub mod boc;
 pub mod evolution;
 pub mod fk_table;
 pub mod grid;
 pub mod import_only_subgrid;
+pub mod pids;
 pub mod subgrid;
 
 /// PyO3 Python module that contains all exposed classes from Rust.
-///
-/// NOTE: this name has to match the one in Cargo.toml 'lib.name'
 #[pymodule]
 fn pineappl(m: &Bound<'_, PyModule>) -> PyResult<()> {
-    m.add_class::<bin::PyBinRemapper>()?;
-    m.add_class::<evolution::PyEvolveInfo>()?;
-    m.add_class::<grid::PyGrid>()?;
-    m.add_class::<grid::PyOrder>()?;
-    m.add_class::<grid::PyOperatorSliceInfo>()?;
-    m.add_class::<grid::PyPidBasis>()?;
-    m.add_class::<channel::PyChannel>()?;
-    m.add_class::<import_only_subgrid::PyImportOnlySubgridV1>()?;
-    m.add_class::<import_only_subgrid::PyImportOnlySubgridV2>()?;
-    m.add_class::<fk_table::PyFkTable>()?;
-    m.add_class::<fk_table::PyFkAssumptions>()?;
-    m.add_class::<subgrid::PySubgridEnum>()?;
-    m.add_class::<subgrid::PySubgridParams>()?;
-    m.add_class::<subgrid::PyMu2>()?;
+    bin::register(m)?;
+    boc::register(m)?;
+    grid::register(m)?;
+    import_only_subgrid::register(m)?;
+    evolution::register(m)?;
+    fk_table::register(m)?;
+    pids::register(m)?;
+    subgrid::register(m)?;
     m.add("version", env!("CARGO_PKG_VERSION"))?;
 
     Ok(())