+++ CHANGELOG.md
@@ -12,6 +12,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - added new method `Grid::delete_orders` and the corresponding switch
   `--delete-orders` in the subcommand `write` of the CLI
 
+### Changed
+
+- Python API: dropped top-level Python interface layer
+- Python API: renamed `lumi` to `channel` in PyO3 Python interface. This
+  concerns 1) the argument names of `convolute_with_one` and similar functions;
+  2) the module `pineappl.lumi` was moved to `pineappl.boc`; 3) the class
+  `LumiEntry` was renamed to `Channel`
+- Python API: `.into()` needs to be explicitly called on subgrids when calling
+  `pineappl.grid.set_subgrid()`
+- Python API: replaced `pineappl.grid.PyPidBasis` with
+  `pineappl.evolution.PidBasis`
+- Python API: replaced `pineappl.grid.PyOperatorSliceInfo` with
+  `pineappl.evolution.OperatorSliceInfo`
+- Python API: drop all `Py` prefixes, for instance `PyEvolveInfo` was renamed
+  to `EvolveInfo`
+
+### Removed
+
+- Python API: removed `pineappl.grid.Grid.create()` and
+  `pineappl.fk_table.FkTable.from_grid()` methods; use the constructors
+  of the respective class instead
+
 ## [0.8.2] - 22/07/2024
 
 ### Changed